[tinker] Move TinkerLogInlineFence to loader package so that it can be removed from patched dex automatically.
